Struct binjs_meta::export::TypeDeanonymizer

pub struct TypeDeanonymizer { /* fields omitted */ }

A tool designed to replace all anonymous types in a specification of the language by explicitly named types.

Consider the following mini-specifications for JSON:

interface Value {
    attribute (Object or String or Number or Array or Boolean)? value;
}
interface Object {
    attribute FrozenArray<Property> properties;
}
interface Property {
    attribute DOMString name;
    attribute Value value;
}
interface Array {
    attribute FrozenArray<Value?> items;
}
// ... Skipping definitions of String, Number, Boolean

The deanonymizer will rewrite them as follows:

interface Value { // Deanonymized optional sum
    attribute OptionalObjectOrStringOrNumberOrArrayOrBoolean value;
}
interface Object { // Deanonymized list
    attribute ListOfProperty properties;
}
interface Property { // No change
    attribute DOMString name;
    attribute Value value;
}
interface Array { // Deanonymized list of options
    attribute ListOfOptionalValue items;
}
// ... Skipping definitions of String, Number, Boolean

typedef ObjectOrStringOrNumberOrArrayOrBoolean? OptionalObjectOrStringOrNumberOrArrayOrBoolean;
typedef (Object
         or String
         or Number
         or Array
         or Boolean)
         ObjectOrStringOrNumberOrArrayOrBoolean;
typedef FrozenArray<Property> ListOfProperty;
typedef FrozenArray<OptionalValue> ListOfOptionalValue;
typedef Value? Optionalvalue;

This deanonymization lets us cleanly define intermediate data structures and/or parsers implementing the webidl specification.

Methods

impl TypeDeanonymizer

pub fn new(spec: &Spec) -> Self

Create an empty TypeDeanonymizer.

pub fn into_spec(self, options: SpecOptions) -> Spec

Convert into a new specification.

pub fn get_node_name(&self, name: &str) -> Option<NodeName>

If name is the name of a (deanonymized) type, return the corresponding type.