Enum binjs_meta::spec::TypeSpec

pub enum TypeSpec {
    Array {
        contents: Box<Type>,
        supports_empty: bool,
    },
    NamedType(NodeName),
    TypeSum(TypeSum),
    Boolean,
    String,
    Number,
    Offset,
    Void,
}

The contents of a type, typically that of a field.

Note that we generally use Type, to represent the fact that some fields accept null while others do not.

Variants
Array

An array of values of the same type.

Fields of Array

contents: Box<Type>

The type of values in the array.

supports_empty: bool

If true, the array may be empty.

NamedType(NodeName)TypeSum(TypeSum)Boolean

A boolean.

String

A string.

Number

A number, as per JavaScript specifications.

Offset

A number of bytes in the binary file.

This spec is used only internally, as a hidden field injected by deanonymization, to represent Skippable nodes.

Void

Nothing.

For the moment, this spec is used only internally.

Methods

impl TypeSpec

pub fn array(self) -> Type

pub fn non_empty_array(self) -> Type

pub fn optional(self) -> Option<Type>

pub fn required(self) -> Type

pub fn typenames<'a>(&'a self) -> HashSet<&'a NodeName>

pub fn get_primitive(&self, spec: &Spec) -> Option<IsNullable<Primitive>>