Struct binjs_meta::spec::Interface

pub struct Interface { /* fields omitted */ }

Representation of an interface in a grammar declaration.

Interfaces represent nodes in the AST. Each interface has a name, a type, defines properties (also known as attribute in webidl) which hold values. Interfaces may also have meta-properties, such as their skippability.

Methods

impl Interface

pub fn contents(&self) -> &Obj

Returns the full list of fields for this structure. This method is in charge of:

ensuring that the fields of parent structures are properly accounted for;

disregarding ignored fields (i.e. position, type);

disregarding fields with a single possible value.

pub fn name(&self) -> &NodeName

Returns the name of the interface.

pub fn spec(&self) -> TypeSpec

Returns a type specification for this interface.

The result is a NamedType with this interface's name.

pub fn type_(&self) -> Type

pub fn get_field_by_name(&self, name: &FieldName) -> Option<&Field>

pub fn is_skippable(&self) -> bool

true if parsers should have the ability to skip instances of this interface.